+++ src/utils/actions.ts
@@ -7,6 +7,7 @@ import taskJSON from "./data/task.json";
 import weaponsJSON from "./data/weapons.json";
 import optionsJSON from "./data/options.json";
 import targetTypesJSON from "./data/targetTypes.json";
+import availableActionsJSON from "./data/availableActions.json";
 import { type TreeSelectOption, type SelectOption } from "naive-ui";
 import type { TActionType, TConstsList, TGroupTypes, TUnitType, TUpperLevelTasks } from "../types";
 import type { TEnrouteTask, TOptionName, TPerformCommand, TPerformTask } from "./consts";
@@ -22,6 +23,27 @@ type TAutoActions = Record<
   >
 >;
 
+const commands = commandsJSON as Record<string, TConstsList<TPerformCommand>>;
+const enrouteTask = enrouteJSON as Record<string, TConstsList<TEnrouteTask>>;
+const performTask = taskJSON as Record<string, TConstsList<TPerformTask>>;
+const options = optionsJSON as Record<number, TConstsList<TOptionName>>;
+const autoActions = autoJSON as TAutoActions;
+const targetTypes = targetTypesJSON as Record<
+  "air" | "airDefense" | "airplanes" | "all" | "ground" | "missles" | "naval",
+  TreeSelectOption
+>;
+const availableWeapons = weaponsJSON as Record<
+  TGroupTypes,
+  Record<TUpperLevelTasks, SelectOption[]>
+>;
+const availableActions = availableActionsJSON as Record<
+  TUnitType,
+  Record<
+    TActionType,
+    Record<TUpperLevelTasks, Array<TPerformTask | TEnrouteTask | TPerformCommand | TOptionName>>
+  >
+>;
+
 export const getFormation = (unitType: Exclude<TUnitType, "ship" | "vehicle">) => {
   if (unitType === "helicopter") {
     return RotaryWingJSON as TreeSelectOption[];
@@ -29,15 +51,49 @@ export const getFormation = (unitType: Exclude<TUnitType, "ship" | "vehicle">) =
   return FixedWingJSON as TreeSelectOption[];
 };
 
-const commands = commandsJSON as Record<string, TConstsList<TPerformCommand>>;
-const enrouteTask = enrouteJSON as Record<string, TConstsList<TEnrouteTask>>;
-const performTask = taskJSON as Record<string, TConstsList<TPerformTask>>;
-const options = optionsJSON as Record<number, TConstsList<TOptionName>>;
-const autoActions = autoJSON as TAutoActions;
-const targetTypes = targetTypesJSON as Record<string, TreeSelectOption>;
-const availableWeapons = weaponsJSON as Record<
-  TGroupTypes,
-  Record<TUpperLevelTasks, SelectOption[]>
->;
+const targetTypesByTask = {
+  all: {
+    SEAD: targetTypes.airDefense,
+    "Antiship Strike": targetTypes.naval,
+    CAS: targetTypes.ground,
+    AFAC: targetTypes.ground,
+    CAP: { ...targetTypes.air, ...targetTypes.missles },
+    Escort: targetTypes.air,
+    "Fighter Sweep": targetTypes.airplanes,
+    All: targetTypes.all,
+  },
+  helicopter: {
+    Escort: targetTypes.ground,
+  },
+};
+
+export const getTargetTypes = (unitType: TUnitType, task: TUpperLevelTasks) => {
+  if (Object.keys(targetTypesByTask.all).includes(task)) {
+    return targetTypesByTask.all[task as keyof typeof targetTypesByTask.all];
+  } else if (unitType === "helicopter" && task === "Escort") {
+    return targetTypesByTask.helicopter[task as keyof typeof targetTypesByTask.helicopter];
+  } else {
+    return {};
+  }
+};
+
+export const getAvailableActions = (
+  unitType: TUnitType,
+  actionType: TActionType,
+  task: TUpperLevelTasks,
+) => {
+  const actions = availableActions[unitType][actionType][task];
+
+  return actions.map((action) => {
+    if (actionType === "task") {
+      return performTask[action];
+    } else if (actionType === "enrouteTask") {
+      return enrouteTask[action];
+    } else if (actionType === "commands") {
+      return commands[action];
+    }
+    return options[action as TOptionName];
+  });
+};
 
 export { commands, enrouteTask, performTask, options, autoActions, availableWeapons, targetTypes };